+++ kernel/interface/fs/fat32/fat32.h
@@ -5,6 +5,6 @@
 #include <interface/fio.h>
 #include <fs/fat32.h>
 
-int Fat32IRead(const char *path, enum File_Mode mode, char *buf, unsigned long size, unsigned long offset, void *key);
+long Fat32IRead(const char *path, enum File_Mode mode, char *buf, unsigned long size, unsigned long offset, void *key);
 
 FilesystemInterface GetFat32Interface(FatFS *fatfs);
