+++ kernel/interface/fs/voidelle/read.c
@@ -2,7 +2,7 @@
 #include <interface/fio.h>
 #include <interface/errno.h>
 
-int VoidelleIRead(const char *path, enum File_Mode mode, char *buf, unsigned long size, unsigned long offset, void *key)
+long VoidelleIRead(const char *path, enum File_Mode mode, char *buf, unsigned long size, unsigned long offset, void *key)
 {
     Voidom voidom = *(Voidom *)key;
